It appears that a sizeable fraction of the running time during the construction of xor filters is caused by expensive cache misses. In such cases, one can see to group together elements that are subject to cache misses (due to random access)

It looks like one little change can boost the speed by about 15% for large sets, and over 5% for modest sets. I would not normally bother with such a small changes, but here, very little change is involved. The readability of the code is not impaired.

I think that because of the way I went about it, the memory usage might be increased, but this could be tweaked after more benchmarks. 


For modest sets:


Before....

```
$ ./bulk-insert-and-query.exe 10000000 0
                                     million    find    find    find    find    find                       optimal  wasted million
                                    adds/sec      0%     25%     50%     75%    100%        ε  bits/item  bits/item   space    keys
                            Xor8        7.39   47.58   47.60   47.61   47.58   47.56  0.3916%       9.84       8.00   23.1%    10.0
```
                                    

After....

```
$ ./bulk-insert-and-query.exe 10000000 0
                                     million    find    find    find    find    find                       optimal  wasted million
                                    adds/sec      0%     25%     50%     75%    100%        ε  bits/item  bits/item   space    keys
                            Xor8        8.06   47.53   47.53   47.53   47.56   47.59  0.3923%       9.84       7.99   23.1%    10.0
```


For larger sets...


Before....

```
$ ./bulk-insert-and-query.exe 100000000 0
                                     million    find    find    find    find    find                       optimal  wasted million
                                    adds/sec      0%     25%     50%     75%    100%        ε  bits/item  bits/item   space    keys
                            Xor8        6.55   37.70   37.69   37.70   37.70   37.70  0.3916%       9.84       8.00   23.1%   100.0
```

After....

```
$ ./bulk-insert-and-query.exe 100000000 0
                                     million    find    find    find    find    find                       optimal  wasted million
                                    adds/sec      0%     25%     50%     75%    100%        ε  bits/item  bits/item   space    keys
                            Xor8        7.46   37.71   37.69   37.70   37.68   37.69  0.3897%       9.84       8.00   22.9%   100.0
```